Fix Year 2016 problem (Bug 1530670)
It will always calculate the GPS year relative to the current system date, 12 years to the past and 4 years to the future. In other words: 0 will be interpreted as 2016 in the years 2012..2027. When run in 2028 (and his GPS is still alive) it will be interpreted as 2032 etc.

Previous situation: The resulting year was always between 2000..2015 because the year field in the device only has 4 bits.